Tons of changes
From discord: Don't have time to test things tonight, but I made a lot of changes and added a new script as well for GPOs. The startup-script will now run all jobs in parallel and wait until all jobs are finished and I've created a script GPOs.ps1 that will download the DoD GPOs then install them, don't know if this works or not, but its there to test if anyone gets up early enough in the morning to test before the competition. Please approve the changes so we can atleast try and test before the competition.